feat(INTERNAL-1370): finish interview on reject
