+++ src/modules/commentMethods.ts
@@ -5,19 +5,24 @@ import { Paths, generatePath } from '../utils/paths';
 import config from '../config';
 import { ErrorWithStatus, idsToIdObjs } from '../utils';
 
+import { interviewMethods } from './interviewMethods';
+import { hireStreamMethods } from './hireStreamMethods';
 import { sendMail } from './nodemailer';
 import { CreateComment, EditComment, DeleteComment } from './commentTypes';
 import { tr } from './modules.i18n';
 import { crewMethods } from './crewMethods';
 import { VacancyStatus } from './crewTypes';
+import { HistorySubject } from './historyEventTypes';
+import { historyEventMethods } from './historyEventMethods';
+import { analyticsEventMethods } from './analyticsEventMethods';
 
-const createComment = async (params: CreateComment): Promise<Comment> => {
-    const { text, userId, target, attachIds, ...restData } = params;
+const createComment = async (params: CreateComment, sessionUserId: number): Promise<Comment> => {
+    const { text, userId, target, attachIds, status } = params;
 
     const data: Prisma.CommentCreateInput = {
-        ...restData,
+        status,
         text,
-        user: { connect: { id: params.userId } },
+        user: { connect: { id: userId } },
         attaches: attachIds ? { connect: idsToIdObjs(attachIds) } : undefined,
     };
 
@@ -27,6 +32,43 @@ const createComment = async (params: CreateComment): Promise<Comment> => {
 
     if ('interviewId' in target) {
         data.interview = { connect: { id: target.interviewId, status: target.status } };
+
+        const previousInterview = await interviewMethods.findWithSections(target.interviewId);
+        const hireStream = await hireStreamMethods.getById(previousInterview.hireStreamId);
+
+        const updatedInterview = await interviewMethods.update({
+            status: target.status,
+            interviewId: target.interviewId,
+        });
+
+        const commonHistoryFields = {
+            userId: sessionUserId,
+            subject: HistorySubject.INTERVIEW,
+            subjectId: target.interviewId,
+        };
+
+        if (status && previousInterview.status !== target.status) {
+            await historyEventMethods.create({
+                ...commonHistoryFields,
+                action: 'set_status',
+                before: previousInterview.status,
+                after: target.status,
+            });
+        }
+
+        if (target.status === 'HIRED' || target.status === 'REJECTED') {
+            await analyticsEventMethods.createEvent(
+                {
+                    event: 'candidate_finished_interview',
+                    interviewId: updatedInterview.id,
+                    candidateId: previousInterview.candidateId,
+                    hireStream: hireStream.name,
+                    hire: target.status === InterviewStatus.HIRED,
+                    rejectReason: text,
+                },
+                sessionUserId,
+            );
+        }
     }
 
     const comment = await prisma.comment.create({
