+++ src/modules/modules.i18n/en.json
@@ -120,5 +120,6 @@
     "Minimum length is {length}": "",
     "Minimum value is {value}": "",
     "Hire streams events can only be requested for 1 day or week": "",
-    "Valid characters: a-z, _, digits": ""
+    "Valid characters: a-z, _, digits": "",
+    "Rejection following the results of {title} section": ""
 }
