+++ src/trpc/routers/sectionsRouter.ts
@@ -1,8 +1,4 @@
 import { accessMiddlewares } from '../../modules/accessMiddlewares';
-import { analyticsEventMethods } from '../../modules/analyticsEventMethods';
-import { hireStreamMethods } from '../../modules/hireStreamMethods';
-import { interviewMethods } from '../../modules/interviewMethods';
-import { sectionTypeMethods } from '../../modules/sectionTypeMethods';
 import { sectionMethods } from '../../modules/sectionMethods';
 import {
     cancelSectionSchema,
@@ -15,37 +11,10 @@ import { protectedProcedure, router } from '../trpcBackend';
 import { historyEventMethods } from '../../modules/historyEventMethods';
 import { HistorySubject } from '../../modules/historyEventTypes';
 import { prisma } from '../../utils/prisma';
-import { notifyHR } from '../../modules/emailMethods';
-import config from '../../config';
-import { userMethods } from '../../modules/userMethods';
-import { crewMethods } from '../../modules/crewMethods';
-import { idObjsToIds } from '../../utils';
 
 const hireStatusToString = (hire: boolean | null | undefined) => {
     if (hire === undefined || hire === null) return;
     return hire ? 'HIRE' : 'NO_HIRE';
-};
-
-const giveSectionAchievement = async (interviewerId: number) => {
-    if (!config.crew.techUserEmail) return;
-
-    const amountOfInterviewerCompletedSections = await prisma.section.count({
-        where: {
-            interviewers: { some: { id: interviewerId } },
-            feedback: { not: null },
-            sectionType: { giveAchievement: true },
-        },
-    });
-
-    const { email } = await userMethods.find(interviewerId);
-
-    const achievementResult = await crewMethods.giveAchievement({
-        targetUserEmail: email,
-        actingUserEmail: config.crew.techUserEmail,
-        sectionsNumber: amountOfInterviewerCompletedSections,
-    });
-
-    console.log(`Crew achievement response: ${achievementResult}`);
 };
 
 export const sectionsRouter = router({
@@ -81,35 +50,8 @@ export const sectionsRouter = router({
             const { data } = input;
 
             const sectionBeforeUpdate = await sectionMethods.getById(data.sectionId);
-            const createFinishSectionEvent = sectionBeforeUpdate.hire === null && data.hire !== null;
 
             const result = await sectionMethods.update(data, ctx.session.user);
-
-            const previousInterview = await interviewMethods.findWithSections(data.interviewId);
-
-            if (createFinishSectionEvent) {
-                const sectionType = await sectionTypeMethods.getById({ id: result.sectionTypeId });
-                const hireStream = await hireStreamMethods.getById(previousInterview.hireStreamId);
-                await analyticsEventMethods.createEvent(
-                    {
-                        event: 'candidate_finished_section',
-                        candidateId: previousInterview.candidateId,
-                        interviewId: data.interviewId,
-                        interviewerIds: idObjsToIds(result.interviewers),
-                        sectionId: result.id,
-                        sectionType: sectionType.value,
-                        hireStream: hireStream.name,
-                        hire: data.hire ?? result.hire ?? false,
-                        grade: data.grade ?? undefined,
-                    },
-                    ctx.session.user.id,
-                );
-
-                await notifyHR(result.id, data);
-                if (sectionType.giveAchievement) {
-                    await Promise.all(result.interviewers.map(({ id }) => giveSectionAchievement(id)));
-                }
-            }
 
             const commonHistoryFields = {
                 userId: ctx.session.user.id,
